Move Rect out of maps directory for general use
